+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## [v1.1.0](https://github.com/cacing69/cquery/compare/v1.0.0...v1.0.1) (2023-0x-xx)
+## [v1.1.0](https://github.com/cacing69/cquery/compare/v1.0.0...v1.0.1) (2023-09-xx)
 - Add method onContentLoaded(Closure $closure) for submit or go to another page when content loaded
 - Add method eachItem(Closure $closure) to manipulate each item
 - Add method onObtainedResults(Closure $closure) to manipulate result array with closure callback